Implement Geo types (#537)
* Add support for Column<GeoPoint>

* restrict tests to heavydb 6.4

* Add docstring for Point2D

* update test_testing to include geopoint and linestring

* refactor Column<GeoPoint>::getItem

* fix geopoint docstring

* fix point2d docstring

* [wip] Add Column<GeoLineString> and GeoLineString

* update geopoint docstring

* * Add docs
* Update code to simplify some implementations
* flake8

* * Fix error where the same struct was generated for different HeavyDB types
* isor

* rename x -> col; i -> index

* rewrite ColumnGeoLineString::get_item

* Fix CI bug

* fix bug in custom_params usage

* Add support for MultiLineString, [Multi]Polygon, MultiPoint

* add geo methods isNull, size and getItem

* Implement fromCoords and add more tests

* - Expose geo_fromCoords_vec/vec2
- Add geolinestring.fromCoords
- Add geomultilinestring.fromCoords
- Add geopolygon.toCoords
- Use builtin list instead of heavydb Array
- Add ct_to_multilinestring test
- isort

* Add ct_make_multipolygon test

* add ct_to_polygon

* Fix bug in geolinestring

* add ct_make_polygon3 test

* rename toCoords -> to_coords
fromCoords -> from_coords

* Update docs

* skip loading multipoint/multilinestring when HeavyDB version < 6.2

* fixes for python < 3.9

* Adjust PR based on comments made on HeavyDB PR

* mark specific test case as xfail

* Try resolving the exit code 143 issue.

---------

+++ rbc/tests/__init__.py
@@ -1,7 +1,7 @@
 __all__ = ['heavydb_fixture', 'sql_execute']
 
 
-from abc import abstractclassmethod, abstractproperty
+from abc import abstractproperty
 import numpy as np
 import os
 import pytest
@@ -37,12 +37,16 @@ def sql_execute(query):
 
 class _TestTable:
 
-    @abstractclassmethod
+    @classmethod
     def suffix(cls):
-        pass
+        assert cls.__name__.startswith('_')
+        assert cls.__name__.endswith('TestTable')
+        return cls.__name__[1:-len('TextTable')].lower()
 
     @abstractproperty
     def sqltypes(self):
+        # TODO: for readability, sqltypes should return a mapping from
+        # column name -> column type
         pass
 
     @property
@@ -56,6 +60,10 @@ class _TestTable:
     @property
     def table_defn(self):
         return ',\n'.join(f'{n} {t}' for n, t in zip(self.colnames, self.sqltypes))
+
+    @property
+    def require_version(self):
+        return (5, 7, 0)
 
 
 class _DefaultTestTable(_TestTable):
@@ -120,7 +128,7 @@ class _nullTestTable(_DefaultTestTable):
         }
 
 
-class _arrayTestTable(_DefaultTestTable):
+class _arrayTestTable(_TestTable):
 
     @classmethod
     def suffix(cls):
@@ -216,8 +224,176 @@ class _TextTestTable(_TestTable):
         }
 
 
+class _PointTestTable(_TestTable):
+
+    @property
+    def sqltypes(self):
+        return ("POINT",
+                "GEOMETRY(POINT, 4326)",
+                "GEOMETRY(POINT, 4326) ENCODING NONE",
+                "GEOMETRY(POINT, 900913)")
+
+    @property
+    def values(self):
+        return {
+            'p1': ['POINT(1 2)', 'POINT(9 8)', None],
+            'p2': ['POINT(3 4)', 'POINT(7 6)', None],
+            'p3': ['POINT(5 6)', 'POINT(5 4)', None],
+            'p4': ['POINT(7 8)', 'POINT(3 2)', None],
+        }
+
+
+class _LineStringTestTable(_TestTable):
+
+    @property
+    def sqltypes(self):
+        return ("LINESTRING",
+                "GEOMETRY(LINESTRING, 4326) ENCODING NONE",
+                "GEOMETRY(LINESTRING, 4326) ENCODING NONE",
+                "GEOMETRY(LINESTRING, 900913)")
+
+    @property
+    def values(self):
+        return {
+            'l1': ['LINESTRING(1 2, 3 5)', 'LINESTRING(9 8, 11 11)', None],
+            'l2': ['LINESTRING(3 4, 5 7)', 'LINESTRING(7 6, 9 9)', None],
+            'l3': ['LINESTRING(5 6, 7 9)', 'LINESTRING(5 4, 7 7)', None],
+            'l4': ['LINESTRING(7 8, 9 11)', 'LINESTRING(3 2, 5 5)', None],
+        }
+
+
+class _MultiLineStringTestTable(_TestTable):
+
+    @property
+    def require_version(self):
+        return (6, 2, 0)
+
+    @property
+    def sqltypes(self):
+        return ("MULTILINESTRING",
+                "GEOMETRY(MULTILINESTRING, 4326) ENCODING NONE",
+                "GEOMETRY(MULTILINESTRING, 4326) ENCODING NONE",
+                "GEOMETRY(MULTILINESTRING, 900913)")
+
+    @property
+    def values(self):
+        return {
+            'ml1': ['MULTILINESTRING((1 2,3 4,5 6,7 8,9 10),(2 3,3 4,1 2))',
+                    'MULTILINESTRING((0 0,5 0,5 5,0 5))',
+                    'MULTILINESTRING((1 2,3 4,5 6,7 8,9 10),(3 4,1 2,2 3),(5 6,7 8,9 10))',
+                    None],
+            'ml2': ['MULTILINESTRING((0 0,5 0,5 5,0 5),(2 2, 2 1,1 1,1 2))',
+                    'MULTILINESTRING((0 0,6 0,6 6,0 6))',
+                    'MULTILINESTRING((0 0,5 0,5 5,0 5),(2 2,2 1,1 1,1 2),(0 0,0 1,1 0))',
+                    None],
+            'ml3': ['MULTILINESTRING((0 0,6 0,6 6,0 6),(3 3,3 2,2 2,2 3))',
+                    'MULTILINESTRING((0 0,7 0,7 7,0 7))',
+                    'MULTILINESTRING((0 0,6 0,6 6,0 6),(3 3,3 2,2 2,2 3),(0 0,0 1,1 0))',
+                    None],
+            'ml4': ['MULTILINESTRING((0 0,7 0,7 7,0 7),(4 4,2 4, 2 3,4 2))',
+                    'MULTILINESTRING((0 0,4 0,4 4,0 4))',
+                    'MULTILINESTRING((0 0,7 0,7 7,0 7),(4 4,2 4, 2 3,4 2),(0 0,0 1,1 0))',
+                    None],
+        }
+
+
+class _MultiPointTestTable(_TestTable):
+
+    @property
+    def require_version(self):
+        return (6, 2, 0)
+
+    @property
+    def sqltypes(self):
+        return ("MULTIPOINT",
+                "GEOMETRY(MULTIPOINT, 4326)",  # uses geoint compression
+                "GEOMETRY(MULTIPOINT, 4326) ENCODING NONE",
+                "GEOMETRY(MULTIPOINT, 900913)")
+
+    @property
+    def values(self):
+        return {
+            'mp1': ['MULTIPOINT(1 2, 3 5)', 'MULTIPOINT(9 8, 11 11)', None],
+            'mp2': ['MULTIPOINT(3 4, 5 7)', 'MULTIPOINT(7 6, 9 9)', None],
+            'mp3': ['MULTIPOINT(5 6, 7 9)', 'MULTIPOINT(5 4, 7 7)', None],
+            'mp4': ['MULTIPOINT(7 8, 9 11)', 'MULTIPOINT(3 2, 5 5)', None],
+        }
+
+
+class _PolygonTestTable(_TestTable):
+
+    @property
+    def sqltypes(self):
+        return ("POLYGON",
+                "GEOMETRY(POLYGON, 4326)",  # uses geoint compression
+                "GEOMETRY(POLYGON, 4326) ENCODING NONE",
+                "GEOMETRY(POLYGON, 900913)")
+
+    @property
+    def values(self):
+        return {
+            'p1': ['POLYGON((1 2,3 4,5 6,7 8,9 10),(2 3,3 4,1 2))',
+                   'POLYGON((0 0,5 0,5 5,0 5,0 0))',
+                   'POLYGON((1 2,3 4,5 6,7 8,9 10),(3 4,1 2,2 3),(5 6,7 8,9 10))',
+                   None],
+            'p2': ['POLYGON((0 0,5 0,5 5,0 5,0 0),(2 2, 2 1,1 1,1 2,2 2))',
+                   'POLYGON((0 0,6 0,6 6,0 6,0 0))',
+                   'POLYGON((0 0,5 0,5 5,0 5,0 0),(2 2,2 1,1 1,1 2,2 2),(0 0,0 1,1 0))',
+                   None],
+            'p3': ['POLYGON((0 0,7 0,7 7,0 7,0 0),(4 4,2 4, 2 3,4 2,4 4))',
+                   'POLYGON((0 0,7 0,7 7,0 7,0 0))',
+                   'POLYGON((0 0,6 0,6 6,0 6,0 0),(3 3,3 2,2 2,2 3,3 3),(0 0,0 1,1 0))',
+                   None],
+            'p4': ['POLYGON((0 0,6 0,6 6,0 6,0 0),(3 3,3 2,2 2,2 3,3 3))',
+                   'POLYGON((0 0,4 0,4 4,0 4,0 0))',
+                   'POLYGON((0 0,7 0,7 7,0 7,0 0),(4 4,2 4, 2 3,4 2,4 4),(0 0,0 1,1 0))',
+                   None],
+        }
+
+
+class _MultiPolygonTestTable(_TestTable):
+
+    @property
+    def sqltypes(self):
+        return ("MULTIPOLYGON",
+                "GEOMETRY(MULTIPOLYGON, 4326)",  # uses geoint compression
+                "GEOMETRY(MULTIPOLYGON, 4326) ENCODING NONE",
+                "GEOMETRY(MULTIPOLYGON, 900913)")
+
+    @property
+    def values(self):
+        return {
+            'mp1': ['MULTIPOLYGON(((1 2,3 4,5 6,7 8,9 10),(2 3,3 4,1 2)))',
+                    'MULTIPOLYGON(((0 0,5 0,5 5,0 5,0 0)))',
+                    'MULTIPOLYGON(((1 2,3 4,5 6,7 8,9 10),(3 4,1 2,2 3),(5 6,7 8,9 10)))',
+                    None],
+            'mp2': ['MULTIPOLYGON(((0 0,5 0,5 5,0 5,0 0),(2 2, 2 1,1 1,1 2,2 2)))',
+                    'MULTIPOLYGON(((0 0,6 0,6 6,0 6,0 0)))',
+                    'MULTIPOLYGON(((0 0,5 0,5 5,0 5,0 0),(2 2,2 1,1 1,1 2,2 2),(0 0,0 1,1 0)))',
+                    None],
+            'mp3': ['MULTIPOLYGON(((0 0,6 0,6 6,0 6,0 0),(3 3,3 2,2 2,2 3,3 3)))',
+                    'MULTIPOLYGON(((0 0,7 0,7 7,0 7,0 0)))',
+                    'MULTIPOLYGON(((0 0,6 0,6 6,0 6,0 0),(3 3,3 2,2 2,2 3,3 3),(0 0,0 1,1 0)))',
+                    None],
+            'mp4': ['MULTIPOLYGON(((0 0,7 0,7 7,0 7,0 0),(4 4,2 4, 2 3,4 2,4 4)))',
+                    'MULTIPOLYGON(((0 0,4 0,4 4,0 4,0 0)))',
+                    'MULTIPOLYGON(((0 0,7 0,7 7,0 7,0 0),(4 4,2 4, 2 3,4 2,4 4),(0 0,0 1,1 0)))',
+                    None],
+        }
+
+
+test_classes = (_DefaultTestTable, _10TestTable, _nullTestTable, _arrayTestTable,
+                _arraynullTestTable, _TextTestTable, _TimestampTestTable,
+                _PointTestTable, _MultiPointTestTable,
+                _LineStringTestTable, _MultiLineStringTestTable,
+                _PolygonTestTable, _MultiPolygonTestTable)
+
+
+test_suffices = [t.suffix() for t in test_classes]
+
+
 def heavydb_fixture(caller_globals, minimal_version=(0, 0),
-                    suffices=['', '10', 'null', 'array', 'arraynull', 'text', 'timestamp'],
+                    suffices=test_suffices,
                     load_test_data=True, debug=False):
     """Usage from a rbc/tests/test_xyz.py file:
 
@@ -257,6 +433,18 @@ def heavydb_fixture(caller_globals, minimal_version=(0, 0),
 
     f'{heavydb.table_name}timestamp' - contains timestamp t9, t6
                                   where 't' prefix is for timestamp.
+
+    f'{heavydb.table_name}point' - contains columns with GeoPoint
+
+    f'{heavydb.table_name}multipoint' - contains columns with GeoMultiPoint
+
+    f'{heavydb.table_name}linestring' - contains columns with LineString
+
+    f'{heavydb.table_name}multilinestring' - contains columns with MultiLineString
+
+    f'{heavydb.table_name}polygon' - contains columns with Polygon
+
+    f'{heavydb.table_name}multipolygon' - contains columns with MultiPolygon
     """
     rbc_heavydb = pytest.importorskip('rbc.heavydb')
     available_version, reason = rbc_heavydb.is_available()
@@ -365,11 +553,14 @@ def heavydb_fixture(caller_globals, minimal_version=(0, 0),
     # DATE, DECIMAL/NUMERIC, GEOMETRY: POINT, LINESTRING, POLYGON,
     # MULTIPOLYGON, See
     # https://docs.heavy.ai/sql/data-definition-ddl/datatypes-and-fixed-encoding
-    for cls in (_DefaultTestTable, _10TestTable, _nullTestTable, _arrayTestTable,
-                _arraynullTestTable, _TextTestTable, _TimestampTestTable):
+    curr_version = Version('.'.join(map(str, m.version[:2])))
+    for cls in test_classes:
         suffix = cls.suffix()
         if suffix in suffices:
             obj = cls()
+            version = Version('.'.join(map(str, obj.require_version)))
+            if version > curr_version:
+                continue
             m.sql_execute(f'DROP TABLE IF EXISTS {table_name}{suffix}')
             m.sql_execute(f'CREATE TABLE IF NOT EXISTS {table_name}{suffix} ({obj.table_defn})')
             m.load_table_columnar(f'{table_name}{suffix}', **obj.values)
